requirements.txt
numpy
matplotlib
timeit
mpl_toolkits.mplot3d 
matplotlib.animation
math